backend/requirements.txt
# Core Framework - Updated for Python 3.13 compatibility
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
# psycopg2-binary==2.9.9  # Optional: Only needed for PostgreSQL
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.20

# Data Validation & Settings - Updated for Python 3.13
pydantic==2.10.4
pydantic-settings==2.8.0
python-dotenv==1.0.1
email-validator==2.1.1

# Caching & Background Tasks
redis==5.2.1
celery==5.3.6

# Testing
pytest==8.2.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.28.1

# Logging & Monitoring
structlog==25.4.0
prometheus-client==0.21.1
psutil==6.1.0

# File Handling
python-magic==0.4.27
Pillow==11.0.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Phase 4: Advanced Features
weasyprint==62.3
openpyxl==3.1.2
xlsxwriter==3.2.1
reportlab==4.4.3
matplotlib==3.10.1
plotly==5.24.1
elasticsearch==8.14.0
icalendar==6.3.1

# Email Management
aiosmtplib==3.0.1
jinja2==3.1.3

# Report Generation
pandas==2.2.3
reportlab==4.4.3
openpyxl==3.1.2

# Utilities
python-dateutil==2.9.0
pytz==2024.2 